Misc cleanup
The netstandard.1.7.Tests file was never compiled before this
